get_ip tests
