+++ backend/api/newsletters/mutations/subscribe_to_newsletter.py
@@ -1,3 +1,4 @@
+from api.utils import get_ip
 import requests
 import logging
 from integrations.flodesk import SubscriptionResult, subscribe
@@ -82,10 +83,3 @@ def subscribe_to_newsletter(
         )
 
     return NewsletterSubscribeResult(status=SubscriptionResult.UNABLE_TO_SUBSCRIBE)
-
-
-def get_ip(request):
-    x_forwarded_for = request.headers.get("x-forwarded-for")
-    if x_forwarded_for:
-        return x_forwarded_for.split(", ")[0]
-    return request.META.get("REMOTE_ADDR")
